+++ logging/lib/sentry_logging.dart
@@ -1,0 +1,3 @@
+library sentry_logging;
+
+export 'src/logging_integration.dart';