+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Feat: Auto Transactions (#643)
+* Feat: Integration for `logging` (#631)
 * Feat: Add logger name to `SentryLogger` and send errors in integrations to the registered logger (#641)
 
 # 6.1.2